feat: Add new public method `openCheckoutWindow` (#61)
This method allows for a pragmatic fix of the popup blocker issue
described in CL-1535.

The client (e.g. tab.js) can now call this method to open the checkout
window right after the confirmation click and _before_ any network
request is made. This way, the window should not be blocked by the
browser.